refactored some code in commands

+++ app.js
@@ -12,7 +12,7 @@ const riot_key = process.env.API_KEY;
 const bot_key = process.env.TOKEN;
 // eslint-disable-next-line no-undef
 const channel_id = process.env.CHANNEL_ID;
-const prefix = '-';
+const prefix = '=';
 
 const Discord = require("discord.js");
 const client = new Discord.Client({intents: ["GUILDS","GUILD_MESSAGES"]});
@@ -34,7 +34,7 @@ client.on("ready", () => {
   console.log(`bot online`)
 })
 
-client.on("messageCreate", message => {
+client.on("messageCreate", message =>  {
   if(!message.content.startsWith(prefix) || message.author.bot || message.channel.name === 'general') return;
 
   // takes message and slices off the prefix, then splits the string into an array with 2 values, the 'add' command and 'summonerName' argument
@@ -43,15 +43,12 @@ client.on("messageCreate", message => {
   const command = args.shift().toLowerCase(); //removes and returns first item from array which will be 'add' 
   const summoner = args.pop().toLowerCase(); //removes and returns last item from array which will be the summoner name 
   
-  if (command === "add"){
-    let exists;
-    players.forEach(e => {
-      if(e.summonerName === summoner){
-      exists = true
-      client.channels.cache.get(channel_id).send("Summoner already added.")
-    }});
-    if (exists) return;
-    
+   if (command === "add"){
+    const channel = client.channels.cache.get(channel_id);
+
+    const existingPlayers = JSON.stringify(players);
+    if(existingPlayers.includes(summoner)) return channel.send("Summoner already added.");
+
     axios({
       method: 'get',
       url: `https://na1.api.riotgames.com/lol/summoner/v4/summoners/by-name/${summoner}?api_key=${riot_key}`,
@@ -75,42 +72,47 @@ client.on("messageCreate", message => {
         // eslint-disable-next-line no-undef
         path.join(__dirname, './data/players.json'),
         JSON.stringify({players}, null, 2)
-    );
+      );
+      channel.send("added summoner")
     })
     .catch(function (error) {
       console.log(error);
+      channel.send("summoner not found")
     });
-    clearInterval(interval);
-    interval();
-    client.channels.cache.get(channel_id).send("added summoner")
+    // clearInterval(interval);
+    // interval();
     return;
   }
 
   if (command === 'remove'){
+    const channel = client.channels.cache.get(channel_id);
+
+    const existingPlayers = JSON.stringify(players)
+    if(!existingPlayers.includes(summoner)) return channel.send("Summoner not found.")
+
     const filteredPlayers = players.filter(selectedPlayer => selectedPlayer.summonerName != summoner);
     players = filteredPlayers;
-    console.log(filteredPlayers);
 
     fs.writeFileSync(
         // eslint-disable-next-line no-undef
         path.join(__dirname, './data/players.json'),
         JSON.stringify({players}, null, 2)
     );
-    clearInterval(interval);
-    interval();
-    client.channels.cache.get(channel_id).send("removed summoner");
+    // clearInterval(interval);
+    // interval();
+    channel.send("removed summoner");
     return;
   }
 });
 
-const test = () => {
-  console.log(dayjs(1645938093697).utc().local().format('lll'))
-};
+// const test = () => {
+//   console.log(dayjs(1645938093697).utc().local().format('lll'))
+// };
 
-const interval = () => {
-  setInterval(test, 1000)
-};
+// const interval = () => {
+//   setInterval(test, 1000)
+// };
 
-interval();
+// interval();
 
 client.login(bot_key)
